Tweaks to packages panel and keyboard shortcut
