+++ packages/web/src/components/keyboard-shortcuts-dialog.tsx
@@ -34,8 +34,8 @@ export default function KeyboardShortcutsDialog({
             <div>
               <h5 className="font-semibold pt-4 pb-2">Global</h5>
               <ShortcutRow keys={['?']} description="show this dialog" />
-              <ShortcutRow keys={['mod', ';']} description="open Srcbook settings" />
-              <ShortcutRow keys={['mod', 'i']} description="open NPM package install modal" />
+              <ShortcutRow keys={['mod', ';']} description="open package.json" />
+              <ShortcutRow keys={['mod', 'i']} description="open npm package install modal" />
               <h5 className="font-semibold pt-6 pb-2">Markdown edit</h5>
               <ShortcutRow keys={['esc']} description="switch back to preview mode" />
               <ShortcutRow keys={['mod', '↵']} description="switch back to preview mode" />
